Se agrega servicio Habitacion

+++ src/main/java/com/SistemaDeReservas/Hotel/repositories/IPisoRepository.java
@@ -3,5 +3,5 @@ package com.SistemaDeReservas.Hotel.repositories;
 import com.SistemaDeReservas.Hotel.models.Piso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPisoRepository  extends JpaRepository<Piso, Integer> {
+public interface IPisoRepository  extends JpaRepository<Piso, Long> {
 }
